+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ### Added
 
+- Support for custom code generators to convert `Symbol` objects to model
+  Python code.
 - Further checks and messages for errors in equations.
 - Added type hints to Python model template.
 - Support for numerical error handling in Fortran implementation, mirroring